python_backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
openai-whisper>=20231117
numpy>=1.24.0
static-ffmpeg>=2.5
